grafana/data: Move getPanelOptionsWithDefaults from core
