+++ packages/grafana-data/src/panel/getPanelOptionsWithDefaults.test.ts
@@ -1,5 +1,3 @@
-import { mockStandardFieldConfigOptions } from 'test/helpers/fieldConfig';
-
 import {
   ConfigOverrideRule,
   FieldColorModeId,
@@ -12,7 +10,9 @@ import {
   StandardOptionConfig,
   ThresholdsMode,
 } from '@grafana/data';
-import { getPanelPlugin } from 'app/features/plugins/__mocks__/pluginMocks';
+
+import { getPanelPlugin } from '../../test/__mocks__/pluginMocks';
+import { mockStandardFieldConfigOptions } from '../../test/helpers/fieldConfig';
 
 import { getPanelOptionsWithDefaults, restoreCustomOverrideRules } from './getPanelOptionsWithDefaults';
