+++ packages/grafana-data/tsconfig.build.json
@@ -1,4 +1,4 @@
 {
-  "exclude": ["dist", "node_modules", "**/*.test.ts*"],
+  "exclude": ["dist", "node_modules", "test", "**/*.test.ts*"],
   "extends": "./tsconfig.json"
 }
