+++ public/app/features/dashboard/components/HelpWizard/HelpWizard.test.tsx
@@ -2,7 +2,7 @@ import { render, screen } from '@testing-library/react';
 import React from 'react';
 
 import { FieldType, getDefaultTimeRange, LoadingState, toDataFrame } from '@grafana/data';
-import { getPanelPlugin } from 'app/features/plugins/__mocks__/pluginMocks';
+import { getPanelPlugin } from '@grafana/data/test/__mocks__/pluginMocks';
 
 import { PanelModel } from '../../state/PanelModel';
