+++ public/app/features/dashboard/components/PanelEditor/OptionsPaneOptions.test.tsx
@@ -12,9 +12,9 @@ import {
   standardFieldConfigEditorRegistry,
   toDataFrame,
 } from '@grafana/data';
+import { getPanelPlugin } from '@grafana/data/test/__mocks__/pluginMocks';
 import { selectors } from '@grafana/e2e-selectors';
 import { getAllOptionEditors, getAllStandardFieldConfigs } from 'app/core/components/OptionsUI/registry';
-import { getPanelPlugin } from 'app/features/plugins/__mocks__/pluginMocks';
 
 import { PanelModel } from '../../state';
 import { createDashboardModelFixture } from '../../state/__fixtures__/dashboardFixtures';
