+++ public/app/features/dashboard/state/DashboardMigrator.test.ts
@@ -1,11 +1,11 @@
 import { each, map } from 'lodash';
 
 import { DataLinkBuiltInVars, MappingType } from '@grafana/data';
+import { getPanelPlugin } from '@grafana/data/test/__mocks__/pluginMocks';
 import { setDataSourceSrv } from '@grafana/runtime';
 import { config } from 'app/core/config';
 import { GRID_CELL_HEIGHT, GRID_CELL_VMARGIN } from 'app/core/constants';
 import { mockDataSource, MockDataSourceSrv } from 'app/features/alerting/unified/mocks';
-import { getPanelPlugin } from 'app/features/plugins/__mocks__/pluginMocks';
 import { MIXED_DATASOURCE_NAME } from 'app/plugins/datasource/mixed/MixedDataSource';
 
 import { VariableHide } from '../../variables/types';
