+++ public/app/features/dashboard/state/PanelModel.test.ts
@@ -10,11 +10,11 @@ import {
   dateTime,
   TimeRange,
 } from '@grafana/data';
+import { getPanelPlugin } from '@grafana/data/test/__mocks__/pluginMocks';
+import { mockStandardFieldConfigOptions } from '@grafana/data/test/helpers/fieldConfig';
 import { setTemplateSrv } from '@grafana/runtime';
 import { queryBuilder } from 'app/features/variables/shared/testing/builders';
 
-import { mockStandardFieldConfigOptions } from '../../../../test/helpers/fieldConfig';
-import { getPanelPlugin } from '../../plugins/__mocks__/pluginMocks';
 import { PanelQueryRunner } from '../../query/state/PanelQueryRunner';
 import { TemplateSrv } from '../../templating/template_srv';
 import { variableAdapters } from '../../variables/adapters';
