+++ public/app/features/datasources/state/buildCategories.test.ts
@@ -1,5 +1,5 @@
 import { DataSourcePluginMeta } from '@grafana/data';
-import { getMockPlugin } from 'app/features/plugins/__mocks__/pluginMocks';
+import { getMockPlugin } from '@grafana/data/test/__mocks__/pluginMocks';
 
 import { buildCategories } from './buildCategories';
