+++ public/app/features/panel/state/actions.test.ts
@@ -1,8 +1,7 @@
-import { mockStandardFieldConfigOptions } from 'test/helpers/fieldConfig';
-
 import { standardEditorsRegistry, standardFieldConfigEditorRegistry } from '@grafana/data';
+import { getPanelPlugin } from '@grafana/data/test/__mocks__/pluginMocks';
+import { mockStandardFieldConfigOptions } from '@grafana/data/test/helpers/fieldConfig';
 import { PanelModel } from 'app/features/dashboard/state';
-import { getPanelPlugin } from 'app/features/plugins/__mocks__/pluginMocks';
 import { panelPluginLoaded } from 'app/features/plugins/admin/state/actions';
 
 import { thunkTester } from '../../../../test/core/thunk/thunkTester';
