+++ public/app/features/variables/datasource/reducer.test.ts
@@ -1,9 +1,9 @@
 import { cloneDeep } from 'lodash';
 
 import { DataSourceInstanceSettings } from '@grafana/data';
+import { getMockPlugins } from '@grafana/data/test/__mocks__/pluginMocks';
 
 import { reducerTester } from '../../../../test/core/redux/reducerTester';
-import { getMockPlugins } from '../../plugins/__mocks__/pluginMocks';
 import { getDataSourceInstanceSetting } from '../shared/testing/helpers';
 import { getVariableTestContext } from '../state/helpers';
 import { VariablesState } from '../state/types';
